Rename the switch for dora
### What changes are proposed in this pull request?

In Dora code we used to use `DORA_CLIENT_READ_LOCATION_POLICY_ENABLED` to determine if Dora is enabled. This property name is not intuitive. Since this has effectively become the main switch, rename it so people know what they are doing.

Note that we are eventually removing all non-Dora code in this branch so this property will become useless by then. But before that, we will rely on it.

			pr-link: Alluxio/alluxio#17651
			change-id: cid-c422df93ecf9a613add334f51841ed9e4441c12e