Create UTs to listing objects of S3 proxy
### What changes are proposed in this pull request?

A new file `ListStatusTest.java` was created to do UTs about  S3 proxy's listing-objects functions.

### Why are the changes needed?
The formely UTs of S3 proxy 'S3ClientRestApiTest'  is redundant and unreadable. 
Most of UTs of S3 proxy is not suitable for alluxio-dora version. 
So I rebuild UTs of S3 proxy to make UTs concise and correct.

### Does this PR introduce any user facing changes?

no.

			pr-link: Alluxio/alluxio#17700
			change-id: cid-5ee1c6c9541ccfadcfbfa4aea1453d5bbe257541