Use a more efficient way to store content hash
### What changes are proposed in this pull request?

Instead of comparing the fingerprint, we compare the fields in the file info object directly, to make it more efficient.

### Why are the changes needed?

In https://github.com/Alluxio/alluxio/pull/17458, we introduced a mechanism that populates fingerprint and persists that into metastore so that we can skip invalidating unnecessary page caches when the metadata is updated. 

Such mechanism is protected under a property key as it does not perform good. Now after the discussion with @dbw9580 , we came up with a more efficient way where we just leverage the fields in the file info object and we believe we are able to get rid of such switch. 

### Does this PR introduce any user facing changes?

N/A
			pr-link: Alluxio/alluxio#17725
			change-id: cid-c0e7898b674cdfc5094e5eb1ee996b3a580c3bf3