Define configurability of update checker
define new property for configurability of update checker, to be used in conjunction with the update checker enabled property. previously the enabled property has its default value set as a build time constant, but the value could be overridden if set in configuration. now the configurability value, which is set at build time, determines if the enabled property can be configured, and if it is, uses its corresponding value. this pattern allows us to lock the enabled value at build time such that a published jar cannot override this enabled property.

			pr-link: Alluxio/alluxio#17738
			change-id: cid-5e07f324674ebf6ebd395e6a4c297655dd633bce