Change to use the new way to get the ETag of the file
### What changes are proposed in this pull request?

Using the content hash of the file to be the ETag in the s3 api. Add a property key to decide whether to get real content hash when loading file meta from ufs.

### Why are the changes needed?
Now we no longer use Xattr to store the etag of the file. And alluxio adds a new option to get real content hash.

### Does this PR introduce any user facing changes?
No

			pr-link: Alluxio/alluxio#17742
			change-id: cid-b91ae9871b9408c5e111668bebb5a398852c229c