[SMALLFIX] Use field property for 'AbstractCmdRunner'
### What changes are proposed in this pull request?

Use field property 'mActiveJobs' for 'AbstractCmdRunner' instead of
the default value DEFAULT_ACTIVE_JOBS which is hardcoded with 3000.

### Why are the changes needed?

Currently, the number of active jobs are limited at 3000 which is not
enough when we have lots of loading tasks on huge cluster. We will add
command line args to allow user pass the limit of active jobs
(set the value of 'mActiveJobs') the  in the future.

### Does this PR introduce any user facing changes?

N/A