Create UTs of putting bucket for S3 proxy with S3 mock
### What changes are proposed in this pull request?

A new file `CreateBucketTest.java` was created to run putting bucket UTs  for S3 proxy based on Dora version.

### Why are the changes needed?

The original s3 API unit tests were very messy and irregular. It is difficult for us to confirm which testing scenarios we have correctly covered. We want to set up s3mock as the UFS and reorganize the unit test for S3 proxy based on Dora version.

### Does this PR introduce any user facing changes?

none.

			pr-link: Alluxio/alluxio#17732
			change-id: cid-71c55b0633c197d2effc919668cb7b51d5b32648