Remove shading for alluxio-test module
### What changes are proposed in this pull request?

Stop shading alluxio-test.

### Why are the changes needed?

Shading a module is slow and error-prone. The shading of the `alluxio-test` module was previously introduced in https://github.com/Alluxio/alluxio/pull/7765 for journal-related tests. There is no longer a need to shade this module.

As part of this PR, `listStatusRootTest` (introduced in https://github.com/Alluxio/alluxio/pull/15361/) has been moved to the S3 specific tests. This move was made because the test failed after the shading was removed on my local MacOS (although it passes CI on Linux), likely due to an unusual setting of the root directory on MacOS. I am unsure why removing shading on `alluxio-test` also impacts `alluxio-integration-tools-validation`. However, based on the description of https://github.com/Alluxio/alluxio/pull/15361, it does seem to fit better in the S3 specific tests.

### Does this PR introduce any user facing changes?

n/a
			pr-link: Alluxio/alluxio#17798
			change-id: cid-e7570b42d9dd096af93053f84a3be226b66f9120