Remove BaseFileSystem
### What changes are proposed in this pull request?

1. Remove all creation and usage of `BaseFileSystem`
2. There are some job-related APIs in `BaseFileSystem`, which are still used in 3.0. Moved those to `UfsBaseFileSystem`.
3. Removed direct users of the `BaseFileSystem`, including `MountCommand/UnmountCommand/UpdateMountCommand`. The `mount()` API on `FileSystem` will be gone.

### Why are the changes needed?

`BaseFileSystem` is the main entrance of the 2.x File/Block API on the client side. It links file I/O to AlluxioFileInStream and AlluxioFileOutStream. It also routes metadata RPCs to the master.

In 3.0, we introduced `UfsBaseFileSystem` and `DoraCacheFileSystem` (together with some others). We have switched to the new code path in https://github.com/Alluxio/alluxio/pull/17612, we remove the entrance to the old code path and say goodbye.

### Does this PR introduce any user facing changes?

Please list the user-facing changes introduced by your change, including
  1. change in user-facing APIs
  4. addition or removal of property keys
  5. webui

			pr-link: Alluxio/alluxio#17830
			change-id: cid-db9be604dfd12d97173c088cd1914206139de55b