Set min replicate to 0 and unpin file if we fail to replicate a file
### What changes are proposed in this pull request?

Now we unpin and set min replication to 0 if we failed due to IO related reasons. (such as UFS not available, file does not exist etc).

### Why are the changes needed?
Previously if a replication job fails, it would be rescheduled again and again, causing Denial of Service.


### Does this PR introduce any user facing changes?
Yes, the user would expect certain files unpinned or adjusted if it becomes unavailable. 

			pr-link: Alluxio/alluxio#17865
			change-id: cid-a2cde6da1e7a95fb361157307bd026303c0164ad