Rename existing bash scripts with -bash suffix
this is a no-op change that renames the existing bash scripts with a `-bash` suffix. 
placeholder scripts are added for `bin/alluxio`, `bin/alluxio-start.sh`, and `bin/alluxio-stop.sh` that redirect to their `-bash` script.

this is in preparation for merging https://github.com/Alluxio/alluxio/tree/golangCli to address https://github.com/Alluxio/alluxio/issues/17522
when merging this branch, the `bin/alluxio` script will be replaced completely. the start and stop script will be modified to redirect to the matching `bin/alluxio process` command to preserve some degree of backcompat.
			pr-link: Alluxio/alluxio#17960
			change-id: cid-6594bc3c4e2e5e97f071e3c5a10ab15fd374b546