Replace bash-based bin/ scripts with golang
### What changes are proposed in this pull request?

Addresses https://github.com/Alluxio/alluxio/issues/17522

### Why are the changes needed?

The CLI in the bash format has grown to be a giant monolith of bash code. One of the goals in the new main branch is to properly modularize various portions of the codebase and the set of bash scripts within the bin/ folder is one obvious counterexample to this goal.

### Does this PR introduce any user facing changes?

For developers, using/building the CLI will now require having golang 1.18+ installed. Note it will not affect running `mvn install` as part of the standard java compilation.

`bin/alluxio` script is completely changed from before and is 100% guaranteed to break backwards compatibility. Most previous commands are migrated to follow the new guidelines defined, namely `bin/alluxio <service> <operation>`. The previous `bin/alluxio` script is renamed to `bin/alluxio-bash` temporarily but will not be maintained and eventually be deleted.

`bin/alluxio-start.sh` and `bin/alluxio-stop.sh` redirect to the new `bin/alluxio process` subcommand. Backwards compatibility is preserved when calling these scripts and they should behave similarly as before. The previous version of the 2 scripts are renamed with a `-bash` suffix along with their supporting bash scripts. The supporting bash scripts will eventually be deleted.

List of common commands and their updated form:
- `collectInfo` -> `info collect`
- `format` -> `init format`
- `fsadmin report` -> `info report`
- `getConf` -> `conf get`
- `runTests` -> `exec basicIOTest`
- most `fs` commands are still under `fs`
			pr-link: Alluxio/alluxio#17974
			change-id: cid-a62ccb70b78c0c48deda670d3be8b366a28b1171