[SMALLFIX] Fix DataPercentageinAlluxio always shows 0%
### What changes are proposed in this pull request?
The LsCommand displays DataInAlluxioPercentage, but it always shows 0% which is misleading.
This change changes the DataInAlluxioPercentage column to a fix string FILE in the Lscommand.
### Why are the changes needed?

Please clarify why the changes are needed. For instance,
As mentioned above, it fixes the bug that LsCommand always shows 0% for DataInAlluxio.

### Does this PR introduce any user facing changes?
Yes, it will change the display of LsCommand.

It will change the output of LsCommand for files but not directories. For files, it will be the follow format

```
./alluxio fs ls /file_path
          FileSize     PeristenceState    Timestamp    FILE     FilePath 
```

Please list the user-facing changes introduced by your change, including
 It changes the percentage of data in Alluxio to a fix string FILE as shown above.

			pr-link: Alluxio/alluxio#18009
			change-id: cid-8b50cec273f5c8edbe7fa14640c59bd41a431e42