Remove web ufs
### What changes are proposed in this pull request?

Remove the code of web under filesystem.

### Why are the changes needed?

We will remove some unused UFSes, now is WebUnderFileSystem. Some UFSes are currently in a pending state (such as Kodo). If users continue to require them in the future, we can consider ongoing support; otherwise, they will be removed in the future, too.

### Does this PR introduce any user facing changes?

No

			pr-link: Alluxio/alluxio#18051
			change-id: cid-ad1ebd19a7c8f847a3607b6237523020a1b5d86b