Remove web ufs
### What changes are proposed in this pull request?

Remove the code of web under filesystem.

### Why are the changes needed?

We will remove some unused UFSes, now is WebUnderFileSystem. Some UFSes are currently in a pending state (such as Kodo). If users continue to require them in the future, we can consider ongoing support; otherwise, they will be removed in the future, too.

### Does this PR introduce any user facing changes?

No

			pr-link: Alluxio/alluxio#18058
			change-id: cid-742ed90a086612dbd978102c9464a36be8837973