Fix alluxio-fuse when mounting to /tmp
### What changes are proposed in this pull request?

Fix alluxio-fuse bug with `/tmp` symlink on macOS

On macOS, the directory `/tmp/` is a symbolic link to `/private/tmp/`. 
This caused the `fuse_mounted` function to fail when attempting to 
mount a UFS to `/tmp/mnt`, as the `mount` command outputs the path 
as `/private/tmp/mnt`. 

This commit corrects this issue by resolving the symlink before checking the mount status.


### Why are the changes needed?

Fix this bug which fails to detect if alluxio-fuse successfully mounts a remote UFS

### Does this PR introduce any user facing changes?

NA

			pr-link: Alluxio/alluxio#18093
			change-id: cid-60a42e81b3c5f6ffe3f75b733817f4a58ec047cc