Fix a deadlock in FileSystemContext
### What changes are proposed in this pull request?

Fix a deadlock.

### Why are the changes needed?

On the Alluxio client, there is a monitor lock of `FileSystemContext` to sync individual client-to-master operations.

One application thread can successfully hold the monitor lock (`alluxio.client.file.FileSystemContext`)  in order to get worker list, but blocked by another “lock” `FileSystemContextReinitializer`  (waiting for latch on on-going RPCs down to zero) for acquiring the block master client to really connect to master process  (waiting for other callers involving `FileSystemContextReinitializer` to finish). 

On the other hand, another heartbeat thread "config-hash-master-heartbeat-0" is awaking periodically to sync with the master process to fetch the latest configuration. This thread detected the conf update and thus entered `FileSystemContextReinitializer` (bumping latch) but was blocked by waiting for the monitor lock of  `alluxio.client.file.FileSystemContext` in order to get the master address.

This PR moves `getMasterAddress` outside `reinit` block to avoid holding the `Reinitializer` object and wait for the monitor object of `FileSystemContext`.

### Does this PR introduce any user facing changes?

No
			pr-link: Alluxio/alluxio#18109
			change-id: cid-0fab540f7dbef3661acb44e7472425add163b019