Fix the array out of bound exception in presto
### What changes are proposed in this pull request?
Fix the array out of bound exception in presto

### Why are the changes needed?

Because we saw earlier that the LocalCacheManager swallows all exceptions, if there is an error accessing one of the cached files, the exception will be ignored. However, the offset may have changed during the failed read attempt in LocalCacheManager. When LocalCacheManager returns -1 on error and tries to reread from the lower layer storage, the offset could be out of bounds.

Regarding why this issue only appears in versions 2.9.3 and later, it's because the offset became a member variable of our target buffer starting from 2.9.3. In earlier versions it was always a local variable, so there was no compounding of offset errors.

### Does this PR introduce any user facing changes?

no

			pr-link: Alluxio/alluxio#18098
			change-id: cid-c7b949d9146847b14ba672df56e21bc4bc5ad705