Create DoraWorkerServiceHandlerTest.java
What changes are proposed in this pull request?
I have created a test and create a liststatus test for its function.

Why are the changes needed?
Please clarify why the changes are needed. For instance,

add a unit test for DoraWorkerClientServiceHandler.
Does this PR introduce any user facing changes?
No.
			pr-link: Alluxio/alluxio#18059
			change-id: cid-b82706a4419700f017584f3e5579d2ef3410aeb3