Enable config the virtual nodes num per node
The configuration `alluxio.user.consistent.hash.virtual.node.count` is for total virtual nodes of all workers. We need to adjust the configuration according to the number of worker nodes. However, it is difficult to determine how many virtual nodes are enough to avoid the data skew issue.

Based on this background, a better way to solve the data skew issue is to allow user to set the virtual nodes number for a worker node, but not the total virtual nodes of all worker nodes. This is what this PR does.
			pr-link: Alluxio/alluxio#18179
			change-id: cid-fbb4c9ebe13f5c0cce4ac08f9fe1789d97cdf209