Add new worker identity
### What changes are proposed in this pull request?

1. Add an abstract representation of a worker's identity.
The new identity is the same as a `long` based identity that used to be generated by the master. It decouples the concept of an identity and the actual implementation of an identity. It also includes a version field that allows backward compatible evolution of the worker identifier.

2. A new implementation for worker identities that uses UUID.
A worker will generate its own identity instead of acquiring one from the master, unless the user explicitly specifies one in the configuration. The generated identity is then persisted in the worker's working directory, and reloaded at the next worker restart.

### Why are the changes needed?

For Dora workers, the identities are not generated by a centralized service, but by workers themselves. Therefore, we need something that provides randomness and avoids collision between workers, e.g. a UUID. The V1 parser implementation included in this PR uses a UUID as the worker identity.

### Does this PR introduce any user facing changes?

Yes, a new configuration item `alluxio.worker.identity.uuid` is added that allows a user to specify the worker identity.

			pr-link: Alluxio/alluxio#17820
			change-id: cid-3200e78fa24be0126d4d49370f0a266f8c766e12