Modify StressWorkerBench: calculate elapsed time in nanoseconds
For now StressWorkerBench uses milliseconds as the smallest unit of recording. In random read, if file is very small, it is possible to have a duration <1ms. Since instant throughput = bytes read / duration, it will lead to a divide by zero error.
This PR replace millisecond-level record of a file read with a nanosecond-level approach. It also makes the following updates on StressWorkerBench:
- record file read time within `applyOperation()` function, instead of before entering or after exiting it
- wrap metrics about the `applyOperation()` function to a private class, instead of returning bytes read only
- remove getter and setters for WorkerBenchDataPoint, since all elements inside is public
- standardize the unit of output throughput, now unit is MB/s
			pr-link: Alluxio/alluxio#18175
			change-id: cid-a606ca5812f75d8116ecaf7da45da9dedd78ab55