Support StressWorkerBench using consistent hash policy
### What changes are proposed in this pull request?
This PR adds a `--mode` option for StressWorkerBench. This option allows user to choose from a range of file read policies. Possible option values are `hash`(default) and `local-only`.

If use `hash`, the config `alluxio.client.file.dora.ConsistentHashPolicy` is set to `alluxio.client.file.dora.ConsistentHashPolicy`.

If use `local-only`, that config is set to `alluxio.client.file.dora.LocalWorkerPolicy`.

The benchmark will parse this option, and choose the right policy. Also, it will print a log about the policy the user is using in both the benchmark and the policy factory.

### Why are the changes needed?
In previous versions, the default policy is `local-only`.

However, in Alluxio, the default policy is `hash`, using `alluxio.client.file.dora.ConsistentHashPolicy`.

If user wants to use the `local-only` policy, add `--mode local-only` to the end of the benchmark command.
			pr-link: Alluxio/alluxio#18246
			change-id: cid-29676a12710b9c815f3e55f84eb5e4226fea9ad7