Fix UT readMultipleFiles in LocalCacheFileInStreamTest
### What changes are proposed in this pull request?

Added a modification to the code in the LocalCacheFileInStreamTest class.Specifically, the code change addresses an issue with the readMultipleFiles unit test in the LocalCacheFileInStreamTest.
The proposed change to fix this issue is to add the setUfsPath method within the generateURIStatus method and set it to be identical to the Path. This ensures that the UfsPath is properly set for each file, allowing the readMultipleFiles test to pass successfully.

### Why are the changes needed?

The changes in this pull request are needed because they address a specific issue with the LocalCacheFileInStreamTest class's readMultipleFiles unit test.
The readMultipleFiles unit test was failing because it relied on the correct calculation of FieldId in Dora, which is based on the UfsPath. However, in this test, the UfsPath for files was not being set and remained empty, leading to all files having the same FieldId after hash calculation.
The absence of a proper UfsPath for files in the test class caused multiple files to share the same FieldId, causing them to read data from a single file instead of their respective files.
To resolve this issue, the proposed solution is to set the UfsPath within the generateURIStatus method to be identical to the Path. This ensures that each file has a distinct and properly set UfsPath, allowing the readMultipleFiles test to pass without affecting other unit tests.

### Does this PR introduce any user facing changes?

Please list the user-facing changes introduced by your change, including
  1. change in user-facing APIs
  2. addition or removal of property keys
  3. webui

			pr-link: Alluxio/alluxio#18168
			change-id: cid-422d5b4eb8bd94b044dbae9181dfc91810e300ed