Fix chaotic chinese characters with golang cli commands
When debugging load command, it is found that `./bin/alluxio` CLI can't process Chinese characters correctly. There must be something wrong with the CLI encoding or decoding. It seems that the Chinese characters are encoded incorrectly.

This happens because the construction of CLI commands does not initialize environmental variables correctly. OS-level environmental variables are ignored, so the program uses a false encoder.

This pull request fixes this problem.
			pr-link: Alluxio/alluxio#18386
			change-id: cid-a15fb106b4a6cd77087cdbba82e426f1ba1e4b4c