Use generic Exception to catch all exceptions in getStatus()
Also print file path in getStatus() exception on client side

### What changes are proposed in this pull request?

Use generic Exception to catch all exceptions in getStatus(), just like that in all other handlers.
File path will be included in log message for getStatus().

### Why are the changes needed?

Sometimes, exceptions are thrown in under file system and/or other third-party libraries.
We don't know the exact type of the specific exceptions. If they are not caught, client will
report unknown exceptions in logs. This is misleading. 

### Does this PR introduce any user facing changes?

Client will show logs with detailed exception information if there is any.
For getStatus(), file path will also be shown in log.

			pr-link: Alluxio/alluxio#18398
			change-id: cid-b6acee3dd0c3e2e03e66bbf195cbda1e95f6bf64