Load content hash as the etag of the object when the UFS is S3
Recode the content hash into XAttr as the ETag when the UFS is S3. When the file is newly created or loaded from UFS, Alluxio will record the ContentHash into metastore so the user can see the ETag attribute.

When the UFS is Object storage, The user can see the ETag of the file persisted in UFS by this way.

No

			pr-link: Alluxio/alluxio#18440
			change-id: cid-1204732240a5e73959b917eb6d9f0c97e05820dc