Add http server port in the worker net address
### What changes are proposed in this pull request?

This PR enables registering the worker's HTTP server's port in the etcd. This helps to find worker's restful APIs from the Python client.

### Why are the changes needed?

Alluxio Python client (e.g. in ML use cases) needs to connect to the worker's REST APIs. But as the http server port isn't included in the worker's information in the etcd, the client fails to find the API endpoint.

### Does this PR introduce any user facing changes?

No.

			pr-link: Alluxio/alluxio#18499
			change-id: cid-1cf7e0bdc7cc0c9702949bc313de5583d9cc2fb8