Add a sleep during every scan round to avoid dead loop.
For big data tasks (eg, spark), might write a lot of files in the final stage. In that case, there will be a lot of write edge locks, which would trigger lockpool evict. All the locks have refcounts greater than 0, which means the locks cannot be evicted during that scan. There is no pause during each scan, and a dead loop accour. The server load grows up. So we add a wait during each scan to avoid this.