Remove distributedMv
This command is dangerous because if the copy half of the operation fails, it may still end up continuing the second half of the operation that deletes the source data. This would result in data loss since the original data would be deleted with no copy created. We will not aim to support this feature in the 2.x line, thus removing the user facing details and entrypoint for this feature.
			pr-link: Alluxio/alluxio#18561
			change-id: cid-dffdf3ebd01be4be043c09b4ea7a4f8562650171