Downgrade AWS SDK to 2.15.79
### What changes are proposed in this pull request?

Downgrade the version of AWS SDK to 2.15.79

### Why are the changes needed?

The current AWS SDK version 2.16.104 depends on Netty 4.1.63-Final, and Alluxio depends on Netty 4.1.52-Final. There is a conflict between the two versions that causes Netty's native epoll library to fail to be loaded. Consequently, the data server on a worker cannot be started on a domain socket which requires epoll, and thus short circuit read cannot be enabled in a Kubernetes deployment.

### Does this PR introduce any user facing changes?

No.

			pr-link: Alluxio/alluxio#18638
			change-id: cid-7a6e4c471a7e50bebb81b456e9c8810a542e86ba